Implemented bookmark through the product card
