+++ prisma/schema.prisma
@@ -17,23 +17,25 @@ datasource db {
 
 // Prisma auth adapter - start
 model User {
-  id            String          @id @default(auto()) @map("_id") @db.ObjectId
-  name          String?
-  email         String?         @unique
-  emailVerified DateTime?
-  image         String?
-  accounts      Account[]
-  sessions      Session[]
+  id                  String            @id @default(auto()) @map("_id") @db.ObjectId
+  name                String?
+  email               String?           @unique
+  emailVerified       DateTime?
+  image               String?
+  accounts            Account[]
+  sessions            Session[]
  
-  createdAt     DateTime        @default(now())
-  updatedAt     DateTime        @updatedAt
+  createdAt           DateTime          @default(now())
+  updatedAt           DateTime          @updatedAt
 
-  username      String          @unique
-  sellerProfile SellerProfile?
-  biography     String?
-  instagram     String?
-  linkedin      String?
-  coverImage    String?
+  username            String            @unique
+  sellerProfile       SellerProfile?
+  biography           String?
+  instagram           String?
+  linkedin            String?
+  coverImage          String?
+
+  bookmarkedProducts  String[]          @default([])
 }
  
 model Account {
@@ -105,7 +107,7 @@ model Product {
 
   status          String          @default("active")
   views           Int             @default(0)
-  likes           Int             @default(0)
+  bookmarks       Int             @default(0)
   
   seller          SellerProfile?  @relation(fields: [sellerId], references: [id], onDelete: SetNull)
   sellerId        String?         @db.ObjectId
