+++ src/app/dashboard/marketplace/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MarketPlaceItems } from "@/components/marketplace/marketplace-items";
+import { MarketPlaceProductCard } from "@/components/marketplace-product-card";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import {
@@ -23,18 +23,22 @@ import { Input } from "@/components/ui/input";
 import { Slider } from "@/components/ui/slider";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useSession } from "@/context/session-context";
+import { useDebounce } from "@/hooks/useDebounce";
+import { Product } from "@/lib/models/Product";
+import { toggleBookmarkListing } from "@/lib/store/marketplace";
 import { categories, Category, Condition, conditions } from "@/types/types";
 import {
   Box,
   Check,
   ChevronDown,
   Filter,
+  Loader2,
   Search,
   ShoppingBag,
   X,
 } from "lucide-react";
 import Link from "next/link";
-import { useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 
 export default function MarketplacePage() {
   const { session } = useSession();
@@ -46,12 +50,53 @@ export default function MarketplacePage() {
     useState<Category>("All Categories");
   const [priceRange, setPriceRange] = useState<[number, number]>([0, 3000]);
   const [condition, setCondition] = useState<Condition>("Any");
+  const [items, setItems] = useState<Product[]>([]);
+  const debouncedSearchQuery = useDebounce(searchQuery, 500);
+  const [loading, setLoading] = useState(true);
+  const [page, setPage] = useState(1);
+  const [totalCount, setTotalCount] = useState(0);
+  const limit = 12;
 
   const emptyQuery = () => {
     setSearchQuery("");
     setSelectedCategory("All Categories");
     setCondition("Any");
     setPriceRange([0, 3000]);
+  };
+
+  const fetchItems = useCallback(async () => {
+    setLoading(true);
+    try {
+      const params = new URLSearchParams();
+      if (debouncedSearchQuery) params.append("search", debouncedSearchQuery);
+      if (selectedCategory) params.append("category", selectedCategory);
+      if (condition) params.append("condition", condition);
+      if (priceRange[0] !== undefined)
+        params.append("minPrice", priceRange[0].toString());
+      if (priceRange[1] !== undefined)
+        params.append("maxPrice", priceRange[1].toString());
+      params.append("page", page.toString());
+      params.append("limit", limit.toString());
+
+      const res = await fetch(`/api/marketplace?${params.toString()}`);
+      if (!res.ok) throw new Error("Failed to fetch products");
+
+      const data = await res.json();
+      setItems(data.products);
+      setTotalCount(data.totalCount);
+    } catch (err) {
+      console.error("Error fetching items", err);
+    } finally {
+      setLoading(false);
+    }
+  }, [condition, priceRange, debouncedSearchQuery, selectedCategory, page]);
+
+  useEffect(() => {
+    fetchItems();
+  }, [fetchItems]);
+
+  const handleBookmarkProduct = async (itemId: string) => {
+    await toggleBookmarkListing(itemId);
   };
 
   return (
@@ -307,13 +352,108 @@ export default function MarketplacePage() {
         )}
       </div>
 
-      <MarketPlaceItems
-        condition={condition}
-        priceRange={priceRange}
-        searchQuery={searchQuery}
-        selectedCategory={selectedCategory}
-        emptyQuery={emptyQuery}
-      />
+      <div className="mb-10">
+        <div>
+          <div className="flex items-center justify-between mb-4">
+            <h2 className="text-xl font-bold">{"All Items"}</h2>
+            <DropdownMenu>
+              <DropdownMenuTrigger asChild>
+                <Button variant="outline" size="sm" className="rounded-xl">
+                  Sort By: Newest
+                  <ChevronDown className="ml-2 h-4 w-4" />
+                </Button>
+              </DropdownMenuTrigger>
+              <DropdownMenuContent align="end">
+                <DropdownMenuItem>Newest</DropdownMenuItem>
+                <DropdownMenuItem>Price: Low to High</DropdownMenuItem>
+                <DropdownMenuItem>Price: High to Low</DropdownMenuItem>
+                <DropdownMenuItem>Most Popular</DropdownMenuItem>
+              </DropdownMenuContent>
+            </DropdownMenu>
+          </div>
+
+          {loading ? (
+            <div className="flex flex-1 justify-center">
+              <Loader2 className="animate-spin" />
+            </div>
+          ) : items.length > 0 ? (
+            <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4">
+              {items.map((item) => (
+                <MarketPlaceProductCard
+                  key={item.id}
+                  item={item}
+                  isBookmarked={
+                    session?.profile.bookmarkedProducts.includes(item.id) ||
+                    false
+                  }
+                  toggleProductBookmark={handleBookmarkProduct}
+                />
+              ))}
+            </div>
+          ) : (
+            <div className="text-center py-12 bg-white dark:bg-gray-900 rounded-xl border border-gray-200 dark:border-gray-800">
+              <div className="mx-auto w-16 h-16 mb-4 rounded-full bg-celestial-blue-100 dark:bg-celestial-blue-900/30 flex items-center justify-center">
+                <Search className="h-8 w-8 text-celestial-blue-500" />
+              </div>
+              <h3 className="text-lg font-semibold mb-2">No items found</h3>
+              <p className="text-gray-600 dark:text-gray-400 max-w-md mx-auto">
+                {
+                  "We couldn't find any items matching your search. Try adjusting your filters or search for something else."
+                }
+              </p>
+              <Button variant="outline" className="mt-4" onClick={emptyQuery}>
+                Clear All Filters
+              </Button>
+            </div>
+          )}
+        </div>
+
+        {/* Pagination */}
+        {items.length > 0 && (
+          <div className="flex justify-center mt-8">
+            <div className="flex items-center gap-1">
+              <Button
+                variant="outline"
+                size="icon"
+                className="rounded-full w-8 h-8"
+                onClick={() => setPage((p) => Math.max(1, p - 1))}
+                disabled={page === 1}
+              >
+                &lt;
+              </Button>
+
+              {[...Array(Math.ceil(totalCount / limit)).keys()].map((_, i) => {
+                const pageNumber = i + 1;
+                return (
+                  <Button
+                    key={pageNumber}
+                    variant="outline"
+                    size="sm"
+                    className={`rounded-full w-8 h-8 ${
+                      page === pageNumber
+                        ? "bg-celestial-blue-500 text-white border-celestial-blue-500"
+                        : ""
+                    }`}
+                    onClick={() => setPage(pageNumber)}
+                  >
+                    {pageNumber}
+                  </Button>
+                );
+              })}
+
+              <Button
+                variant="outline"
+                size="icon"
+                className="rounded-full w-8 h-8"
+                onClick={() => setPage((p) => p + 1)}
+                disabled={page >= Math.ceil(totalCount / limit)}
+              >
+                &gt;
+              </Button>
+            </div>
+          </div>
+        )}
+      </div>
     </main>
   );
 }
