+++ src/app/profile/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { MarketPlaceProductCard } from "@/components/marketplace/marketplace-product-card";
+import { MarketPlaceProductCard } from "@/components/marketplace-product-card";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
