+++ src/lib/models/User.ts
@@ -1,5 +1,5 @@
-import { productSelect } from "@/types/types";
 import { Prisma } from "@prisma/client";
+import { productSelect } from "./Product";
 
 export const userSelect = Prisma.validator<Prisma.UserSelect>()({
   id: true,
@@ -13,6 +13,7 @@ export const userSelect = Prisma.validator<Prisma.UserSelect>()({
   coverImage: true,
   sellerProfile: { include: { products: { select: productSelect } } },
   createdAt: true,
+  bookmarkedProducts: true,
 });
 
 export type User = Prisma.UserGetPayload<{ select: typeof userSelect }>;
