+++ src/types/types.ts
@@ -12,27 +12,6 @@ export type SellerProfile = Prisma.SellerProfileGetPayload<{
   select: typeof sellerProfileSelect;
 }>;
 
-export const productSelect = Prisma.validator<Prisma.ProductSelect>()({
-  id: true,
-  name: true,
-  seller: { select: { user: true, rating: true } },
-  category: true,
-  condition: true,
-  description: true,
-  photos: true,
-  price: true,
-  tags: true,
-  pickupLocation: true,
-  status: true,
-  views: true,
-  createdAt: true,
-  likes: true,
-});
-
-export type Product = Prisma.ProductGetPayload<{
-  select: typeof productSelect;
-}>;
-
 export const categories = [
   "All Categories",
   "Electronics",
